+++ hnn_core/__init__.py
@@ -1,8 +1,8 @@
 from .dipole import simulate_dipole, read_dipole, average_dipoles
 from .drives import drive_event_times
 from .params import Params, read_params
-from .network import Network, read_spikes
-from .cell_response import CellResponse
+from .network import Network
+from .cell_response import CellResponse, read_spikes
 from .pyramidal import L2Pyr, L5Pyr
 from .basket import L2Basket, L5Basket
 from .parallel_backends import MPIBackend, JoblibBackend
