+++ hnn_core/cell_response.py
@@ -4,6 +4,7 @@
 #          Ryan Thorpe <ryan_thorpe@brown.edu>
 #          Mainak Jas <mjas@mgh.harvard.edu>
 
+from glob import glob
 import numpy as np
 
 from .viz import plot_spikes_hist, plot_spikes_raster
@@ -393,3 +394,58 @@ class CellResponse(object):
                         self._spike_times[trial_idx][spike_idx],
                         int(self._spike_gids[trial_idx][spike_idx]),
                         self._spike_types[trial_idx][spike_idx]))
+
+
+def read_spikes(fname, gid_ranges=None):
+    """Read spiking activity from a collection of spike trial files.
+
+    Parameters
+    ----------
+    fname : str
+        Wildcard expression (e.g., '<pathname>/spk_*.txt') of the
+        path to the spike file(s).
+    gid_ranges : dict of lists or range objects | None
+        Dictionary with keys 'evprox1', 'evdist1' etc.
+        containing the range of Cell or input IDs of different
+        cell or input types. If None, each spike file must contain
+        a 3rd column for spike type.
+
+    Returns
+    ----------
+    cell_response : CellResponse
+        An instance of the CellResponse object.
+    """
+
+    spike_times = list()
+    spike_gids = list()
+    spike_types = list()
+    for file in sorted(glob(str(fname))):
+        spike_trial = np.loadtxt(file, dtype=str)
+        if spike_trial.shape[0] > 0:
+            spike_times += [list(spike_trial[:, 0].astype(float))]
+            spike_gids += [list(spike_trial[:, 1].astype(int))]
+
+            # Note that legacy HNN 'spk.txt' files don't contain a 3rd column
+            # for spike type. If reading a legacy version, validate that a
+            # gid_dict is provided.
+            if spike_trial.shape[1] == 3:
+                spike_types += [list(spike_trial[:, 2].astype(str))]
+            else:
+                if gid_ranges is None:
+                    raise ValueError("gid_ranges must be provided if spike "
+                                     "types are unspecified in the "
+                                     "file %s" % (file,))
+                spike_types += [[]]
+        else:
+            spike_times += [[]]
+            spike_gids += [[]]
+            spike_types += [[]]
+
+    cell_response = CellResponse(spike_times=spike_times,
+                                 spike_gids=spike_gids,
+                                 spike_types=spike_types)
+    if gid_ranges is not None:
+        cell_response.update_types(gid_ranges)
+
+    return CellResponse(spike_times=spike_times, spike_gids=spike_gids,
+                        spike_types=spike_types)
